RWA-977 - move executors to own package
