+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/config/executors/CcdCaseEventsDeadLetterQueueExecutor.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.wacaseeventhandler.config;
+package uk.gov.hmcts.reform.wacaseeventhandler.config.executors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import uk.gov.hmcts.reform.wacaseeventhandler.clients.CcdEventConsumer;
+import uk.gov.hmcts.reform.wacaseeventhandler.clients.CcdCaseEventsDeadLetterQueueConsumer;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -13,17 +13,17 @@ import java.util.stream.IntStream;
 
 @Configuration
 @ConditionalOnProperty("azure.servicebus.enableASB")
-public class CcdEventExecutor {
+public class CcdCaseEventsDeadLetterQueueExecutor {
 
     @Value("${azure.servicebus.threads}")
     private int concurrentSessions;
 
     @Autowired
-    private CcdEventConsumer serviceBusTask;
+    private CcdCaseEventsDeadLetterQueueConsumer serviceBusTask;
 
     @Bean
     @SuppressWarnings("PMD.DataflowAnomalyAnalysis")
-    public void createServiceBus() {
+    public void createDeadLetterQueueServiceBus() {
         final ExecutorService executorService = Executors.newFixedThreadPool(
             Integer.valueOf(concurrentSessions));
