+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/config/executors/CcdCaseEventsExecutor.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.wacaseeventhandler.config;
+package uk.gov.hmcts.reform.wacaseeventhandler.config.executors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
