+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/config/executors/CcdEventExecutor.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.wacaseeventhandler.config;
+package uk.gov.hmcts.reform.wacaseeventhandler.config.executors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import uk.gov.hmcts.reform.wacaseeventhandler.clients.CcdCaseEventsDeadLetterQueueConsumer;
+import uk.gov.hmcts.reform.wacaseeventhandler.clients.CcdEventConsumer;
 
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -13,17 +13,17 @@ import java.util.stream.IntStream;
 
 @Configuration
 @ConditionalOnProperty("azure.servicebus.enableASB")
-public class CcdCaseEventsDeadLetterQueueExecutor {
+public class CcdEventExecutor {
 
     @Value("${azure.servicebus.threads}")
     private int concurrentSessions;
 
     @Autowired
-    private CcdCaseEventsDeadLetterQueueConsumer serviceBusTask;
+    private CcdEventConsumer serviceBusTask;
 
     @Bean
     @SuppressWarnings("PMD.DataflowAnomalyAnalysis")
-    public void createDeadLetterQueueServiceBus() {
+    public void createServiceBus() {
         final ExecutorService executorService = Executors.newFixedThreadPool(
             Integer.valueOf(concurrentSessions));
